+++ hf2p_server/source/math/fast_checksum.h
@@ -1,6 +1,5 @@
 #pragma once
-#include <windows.h>
-#include <stdint.h>
+#include "..\cseries\cseries.h"
 
-uint32_t fast_checksum(const void* key, size_t length, uint32_t initval); // hashlittle
-uint32_t fast_checksum_new();
+ulong fast_checksum(const void* key, size_t length, ulong initval); // hashlittle
+ulong fast_checksum_new();
