+++ hf2p_server/source/memory/ring_buffer.h
@@ -1,9 +1,10 @@
 #pragma once
+#include "..\cseries\cseries.h"
 
 class c_ring_buffer
 {
-	int ring_size;
+	long ring_size;
 	void* elements;
-	int element_count;
+	long element_count;
 };
 static_assert(sizeof(c_ring_buffer) == 0xC);
