+++ hf2p_server/source/networking/network_utilities.cpp
@@ -1,5 +1,6 @@
 #include "network_utilities.h"
 #include "network_globals.h"
+#include <timeapi.h>
 
 void network_get_build_identifiers(long* executable_type, long* executable_version, long* compatible_version)
 {
@@ -8,9 +9,9 @@ void network_get_build_identifiers(long* executable_type, long* executable_versi
 	*compatible_version = 11;
 }
 
-uint32_t network_get_time() // non-original name
+ulong network_get_time() // non-original name
 {
-	uint32_t time;
+	ulong time;
 	if (network_time_locked)
 		time = *g_network_locked_time;
 	else
